flask_socketio_celery/requirements.txt
Flask
Flask-SocketIO
eventlet
celery
redis
python-dotenv
